First step of switching over save data.
This step has it that the dynamic puzzle data, static puzzle data and completion data are in single playerprefs instead of several.

Next step is switching code from using data as strings to using the data directly.
