+++ Assets/Prefabs/Managers/SaveData/PuzzleCompletionSaveData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d950bf82a05d041438f28c3722225edb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
