+++ Assets/Prefabs/Managers/SaveData/PuzzleDynamicSaveData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6fc533628efee44418d81bc3fba76a05
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
