+++ Assets/Prefabs/Managers/SaveData/PuzzleStaticSaveData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 295e8ff35b115db41ada620530a87ddb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
